# Text layout perf improvements