--- a/src/Avalonia.Base/Media/TextFormatting/Unicode/Codepoint.cs
+++ b/src/Avalonia.Base/Media/TextFormatting/Unicode/Codepoint.cs
@@ -166,72 +166,7 @@ namespace Avalonia.Media.TextFormatting.Unicode
         /// <param name="index">The index to read at.</param>
         /// <param name="count">The count of character that were read.</param>
         /// <returns></returns>
-        public static Codepoint ReadAt(IReadOnlyList<char> text, int index, out int count)
-        {
-            count = 1;
-
-            if (index >= text.Count)
-            {
-                return ReplacementCodepoint;
-            }
-
-            var code = text[index];
-
-            ushort hi, low;
-
-            //# High surrogate
-            if (0xD800 <= code && code <= 0xDBFF)
-            {
-                hi = code;
-
-                if (index + 1 == text.Count)
-                {
-                    return ReplacementCodepoint;
-                }
-
-                low = text[index + 1];
-
-                if (0xDC00 <= low && low <= 0xDFFF)
-                {
-                    count = 2;
-                    return new Codepoint((uint)((hi - 0xD800) * 0x400 + (low - 0xDC00) + 0x10000));
-                }
-
-                return ReplacementCodepoint;
-            }
-
-            //# Low surrogate
-            if (0xDC00 <= code && code <= 0xDFFF)
-            {
-                if (index == 0)
-                {
-                    return ReplacementCodepoint;
-                }
-
-                hi = text[index - 1];
-
-                low = code;
-
-                if (0xD800 <= hi && hi <= 0xDBFF)
-                {
-                    count = 2;
-                    return new Codepoint((uint)((hi - 0xD800) * 0x400 + (low - 0xDC00) + 0x10000));
-                }
-
-                return ReplacementCodepoint;
-            }
-
-            return new Codepoint(code);
-        }
-        
-        /// <summary>
-        /// Reads the <see cref="Codepoint"/> at specified position.
-        /// </summary>
-        /// <param name="text">The buffer to read from.</param>
-        /// <param name="index">The index to read at.</param>
-        /// <param name="count">The count of character that were read.</param>
-        /// <returns></returns>
-        public static Codepoint ReadAt(CharacterBufferRange text, int index, out int count)
+        public static Codepoint ReadAt(ReadOnlySpan<char> text, int index, out int count)
         {
             count = 1;