--- a/src/Avalonia.Base/Platform/ITextShaperImpl.cs
+++ b/src/Avalonia.Base/Platform/ITextShaperImpl.cs
@@ -1,4 +1,5 @@
-﻿using Avalonia.Media.TextFormatting;
+﻿using System;
+using Avalonia.Media.TextFormatting;
 using Avalonia.Metadata;
 
 namespace Avalonia.Platform
@@ -13,9 +14,8 @@ namespace Avalonia.Platform
         /// Shapes the specified region within the text and returns a shaped buffer.
         /// </summary>
         /// <param name="text">The text buffer.</param>
-        /// <param name="length">The length of text.</param>
         /// <param name="options">Text shaper options to customize the shaping process.</param>
         /// <returns>A shaped glyph run.</returns>
-        ShapedBuffer ShapeText(CharacterBufferReference text, int length, TextShaperOptions options);
+        ShapedBuffer ShapeText(ReadOnlyMemory<char> text, TextShaperOptions options);
     }   
 }