… src/Avalonia.Base/Media/TextFormatting/CharacterBufferRange.cs
… src/Avalonia.Base/Media/TextFormatting/CharacterBufferReference.cs
… src/Avalonia.Base/Media/TextFormatting/TextCharacters.cs
… src/Avalonia.Base/Media/TextFormatting/TextFormatterImpl.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/LineBreakEnumerator.cs
… src/Avalonia.Base/Media/TextFormatting/UnshapedTextRun.cs
… src/Avalonia.Controls/TextBlock.cs
… src/Avalonia.Controls/TextBox.cs
… src/Avalonia.Headless/HeadlessPlatformStubs.cs
… src/Skia/Avalonia.Skia/TextShaperImpl.cs
… src/Windows/Avalonia.Direct2D1/Media/TextShaperImpl.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/BiDiClassTests.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/GraphemeBreakClassTrieGeneratorTests.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/LineBreakEnumuratorTests.cs
… tests/Avalonia.Controls.UnitTests/Presenters/TextPresenter_Tests.cs
… tests/Avalonia.Skia.UnitTests/Media/GlyphRunTests.cs
… tests/Avalonia.Skia.UnitTests/Media/TextFormatting/SingleBufferTextSource.cs
… tests/Avalonia.Skia.UnitTests/Media/TextFormatting/TextFormatterTests.cs
… tests/Avalonia.Skia.UnitTests/Media/TextFormatting/TextLayoutTests.cs
… tests/Avalonia.Skia.UnitTests/Media/TextFormatting/TextLineTests.cs
… tests/Avalonia.Skia.UnitTests/Media/TextFormatting/TextShaperTests.cs
… tests/Avalonia.UnitTests/HarfBuzzTextShaperImpl.cs
… tests/Avalonia.UnitTests/MockTextShaperImpl.cs
… tests/Avalonia.UnitTests/TextTestHelper.cs