## What does the pull request do?
Improves the performance of text layout by using `ReadOnlySpan<T>` and `ReadOnlyMemory<T>` as much as possible. See https://github.com/AvaloniaUI/Avalonia/discussions/9978 for further discussion.

Notable changes:
 - `GlyphRun.Characters` is a `ReadOnlyMemory<char>` instead of a `IReadOnlyList<char>`.
 - `TextRun.Text` is a is a `ReadOnlyMemory<char>` instead of a `CharacterBufferReference`.
 - `CharacterBufferReference` and `CharacterBufferRange` are removed.

## Perf changes:

Current master:
|                Method |       Mean |    Error |   StdDev |    Gen 0 |   Gen 1 | Allocated |
|---------------------- |-----------:|---------:|---------:|---------:|--------:|----------:|
|       BuildTextLayout |   634.2 us |  5.67 us |  5.30 us |   0.9766 |       - |     25 KB |
| BuildEmojisTextLayout | 1,364.1 us | 12.15 us | 11.37 us |        - |       - |     30 KB |
|   BuildManySmallTexts | 4,043.9 us | 21.54 us | 20.15 us | 156.2500 | 78.1250 |  2,563 KB |
|  VirtualizeTextBlocks | 5,088.5 us | 20.43 us | 19.11 us | 148.4375 | 23.4375 |  2,445 KB |

This PR:
|                Method |       Mean |    Error |   StdDev |    Gen 0 |   Gen 1 | Allocated |
|---------------------- |-----------:|---------:|---------:|---------:|--------:|----------:|
|       BuildTextLayout |   355.3 us |  0.95 us |  0.89 us |   0.9766 |       - |     22 KB |
| BuildEmojisTextLayout |   735.4 us |  2.74 us |  2.56 us |   0.9766 |       - |     27 KB |
|   BuildManySmallTexts | 3,267.1 us | 10.25 us |  9.59 us | 144.5313 | 70.3125 |  2,391 KB |
|  VirtualizeTextBlocks | 4,291.4 us | 14.17 us | 12.56 us | 132.8125 | 15.6250 |  2,273 KB |