@Mikolaytis 2023-01-15
carefully merged to Lunacy. Getting nice results:

before:
|                Method |       Mean |     Error |    StdDev |     Gen0 |     Gen1 |   Gen2 |  Allocated |
|---------------------- |-----------:|----------:|----------:|---------:|---------:|-------:|-----------:|
|       BuildTextLayout |   302.2 us |   6.58 us |  18.87 us |   3.9063 |   0.9766 |      - |   25.18 KB |
| BuildEmojisTextLayout |   273.4 us |   5.24 us |  10.70 us |   4.3945 |   1.9531 | 0.4883 |    28.7 KB |
|   BuildManySmallTexts | 5,353.1 us | 106.52 us | 263.28 us | 414.0625 | 203.1250 |      - | 2532.16 KB |

after:
|                Method |       Mean |    Error |    StdDev |     Gen0 |     Gen1 |   Gen2 | Allocated |
|---------------------- |-----------:|---------:|----------:|---------:|---------:|-------:|----------:|
|       BuildTextLayout |   238.2 us |  4.53 us |   4.23 us |   1.9531 |   0.9766 |      - |  12.15 KB |
| BuildEmojisTextLayout |   195.3 us |  3.66 us |   3.59 us |   1.7090 |   0.7324 | 0.2441 |  10.76 KB |
|   BuildManySmallTexts | 4,568.1 us | 90.70 us | 258.78 us | 390.6250 | 195.3125 |      - | 2368.1 KB |

Numbers are bit different from above because I've added some optimisations that not published yet.