@danipen 2023-01-15
I guess we can some perf improvements in AvaloniaEdit using these changes. Thanks @MrJul!